worked on modules
  -from now on we're using Integer with BigInt
  under the hood

  -implemented Add for Homo and a test
  doesn't compile (LOL)
  borrow checker time, another time

+++ src/lib.rs
@@ -1,5 +1,5 @@
 //! Project root
-pub use num_bigint::BigInt;
+pub use num_bigint::BigInt as Integer;
 pub mod error;
 pub mod helpers;
 pub mod limits;
@@ -9,3 +9,5 @@ pub mod poly;
 pub mod products;
 pub mod quotient;
 pub mod sub;
+
+
